✨[Feat]#17 - Report ViewModel 구현
모델을 뷰모델에서 초기화
