+++ Muje/Muje/Modules/MyPage/ViewModels/Community/ReportsHistoryViewModel.swift
@@ -36,17 +36,17 @@ final class ReportsHistoryViewModel {
     // MARK: - 신고 생성 테스트 버튼 로직 (삭제 예정)
     func createReportTestButtonTapped() async {
         // 공고 신고 Model postId: "00DD1837-4416-46BD-8246-50CE26567717"
-        let report = Report(
-            reportId: UUID(),
-            reporterUserId: FirebaseAuthManager.shared.currentUser?.userId ?? "",
-            reportedUserId: "FTMIffTVLdb8GuhHweHJEgAwFqB2",
-            postId: "00DD1837-4416-46BD-8246-50CE26567717",
-            reportType: ReportType.spam.rawValue,
-            reason: "저한테 욕설을 빈번하게 사용했어요",
-            status: ReportStatus.pending.rawValue
-        )
-        
-        let _ = try? await FirestoreManager.shared.create(report)
+//        let report = Report(
+//            reportId: UUID(),
+//            reporterUserId: FirebaseAuthManager.shared.currentUser?.userId ?? "",
+//            reportedUserId: "FTMIffTVLdb8GuhHweHJEgAwFqB2",
+//            postId: "00DD1837-4416-46BD-8246-50CE26567717",
+//            reportType: ReportType.spam.rawValue,
+//            reason: "저한테 욕설을 빈번하게 사용했어요",
+//            status: ReportStatus.pending.rawValue
+//        )
+//        
+//        let _ = try? await FirestoreManager.shared.create(report)
     }
     
 }
